requirements.txt
langid==1.1.6
LanguageDetection==0.1.3
nltk==3.7
pandas==1.4.2
pyarrow==12.0.0
regex==2022.3.15
praw==7.7.0
python-dotenv==1.0.0
